+++ packages/npm/lib/react/radio.js
@@ -58,7 +58,7 @@ var Radio = function (_React$Component) {
           null,
           _react2.default.createElement('input', {
             ref: function ref(el) {
-              _this2.inputElRef = el;
+              _this2.controlEl = el;
             },
             type: 'radio',
             autoFocus: autoFocus,